backend/pyproject.toml
[project]
name = "gideon"
version = "0.1.0"
description = "AI-powered criminal defense discovery platform"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [{ name = "Gideon Contributors" }]
keywords = ["legal", "discovery", "criminal-defense", "rag"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: FastAPI",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

dependencies = [
    # Shared models
    "gideon-shared",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.7",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "pyotp>=2.9",
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    # Vector store
    "qdrant-client>=1.13",
    # RAG / LLM
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-ollama>=0.3",
    "langchain-text-splitters>=0.3",
    # Background jobs
    "celery[redis]>=5.4",
    "psycopg2-binary>=2.9",
    # Document parsing
    "python-multipart>=0.0.18",
    # S3-compatible object storage
    "minio>=7.0",
    # Hashing / crypto
    "cryptography>=44",
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-celery>=0.50b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b0",
    "opentelemetry-exporter-otlp-proto-http>=1.29",
]

[project.optional-dependencies]
monitoring = ["flower>=2.0"]
dev = ["ruff>=0.9", "mypy>=1.14", "types-python-jose>=3.3", "pre-commit>=4.1"]
test = [
    "pytest>=8.3",
    "pytest-bdd>=8.1",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "factory-boy>=3.3",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
test-unit = "pytest -m 'not integration' {args}"
test-integration = "pytest -m integration {args}"
lint = "ruff check . {args}"
format = "ruff format . {args}"
typecheck = "mypy app/ {args}"

[dependency-groups]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "types-python-jose>=3.3",
    "pre-commit>=4.1",
    "pytest>=8.3",
    "pytest-bdd>=8.1",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "factory-boy>=3.3",
    "hatch>=1.16.5",
    "pytest-docker>=3.1",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
]

[tool.uv.sources]
gideon-shared = { workspace = true }

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "S",   # bandit (security)
    "B",   # bugbear
    "SIM", # simplify
    "T20", # print statements
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S104", "S603", "S607"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
bdd_features_base_dir = "tests/features/"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "integration: tests that require a running server or database",
]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true
